WIP: libc-free shake-out — EnvGet/ErrnoOf/Prng + path/Str overloads
Branch is libc-free now: zero <stdlib/stdio/string/errno/math/limits/
stdint/stddef.h> includes anywhere in scanned source, and zero
undefined libc symbols across all test binaries. Saved as a WIP so we
have a recovery point before the next sweep.

Naming pass:
  - SYS_ERRNO(ret) -> ErrnoOf(ret) -- PascalCase static-inline, no
    Sys/SYS_ prefix.
  - SysErrno() -> Errno(); errno_ptr() helper inlined into Errno()
    body (no second symbol to keep in sync).
  - SysGetenv -> EnvGet. On Linux x86_64/aarch64 EnvGet walks our own
    envp captured by _StartLinux.c (new misra_envp + misra_start_c
    helper) -- no libc getenv reference left in the binary.

New: Std/Prng.h
  - Prng8/16/32/64 over xorshift64*, seeded from getrandom (Linux) /
    getentropy (Darwin) / BCryptGenRandom (Windows). State reseeds
    every MISRA_PRNG_RESEED_INTERVAL calls (new meson option
    `prng_reseed_interval`, default 10000, routed through generated
    Config.h).
  - prng_internal holds the ONE program-lifetime static in MisraStdC,
    documented as the explicit exception. State cannot leak: no
    accessor, no out-param.

New: Sys/Dir.h direct-syscall family
  - DirCreate / DirCreateAll (mkdir -p) / DirRemoveAll (rm -rf) via
    MISRA_SYS_mkdir / MISRA_SYS_mkdirat on Linux x86_64/aarch64 and
    Darwin (#136 / #475). Dir.c libc fallbacks deleted -- unsupported
    arches hit `#error` instead of silently binding extern unlink/
    rmdir/mkdir/access.
  - dir_remove_all stack-promotes its per-entry "parent/child" Str via
    StrInitStack(4096) -- the prior HeapAllocator-per-iter is gone.

New: FileRead Str overload + FileOpenTemp
  - read_complete_file + ReadCompleteFile removed; replaced by a
    2-arg FileRead(File*, Str*) overload that reads to EOF into a Str
    via VecReserve chunks (4 KiB block). 5 parser callers (Elf, MachO,
    Pdb, Pe, Http) migrated to FileRead(&f, &str) + move-data-into-
    owner pattern.
  - FileOpenTemp = atomic O_CREAT|O_EXCL|0600 with a 16-hex suffix
    from Prng64. Replaces libc mkstemp.

Path-arg dispatch
  - No MISRA_ZSTR_OF / shared "extract zstr" helper. Each File/Dir/
    parser-file-open macro inlines its own _Generic over `path` ->
    {Str*, const Str*, char*, const char*}. Docs lead with Str*; char*
    is documented as the literal-buffer convenience.
  - APIs touched: FileOpen, FileRead, FileOpenTemp, FileRemove,
    DirRemove, DirCreate, DirCreateAll, DirRemoveAll, DirGetContents,
    FileGetSize, ElfFileOpen, PeFileOpen, MachoFileOpen, PdbFileOpen.

Misc
  - Added Tests/Util/check_no_libc.py + CheckNoLibc test that fails
    the suite on any libc #include in the scanned tree.
  - Misra/Sys/Errno.h: in-tree POSIX errno numeric constants
    (EPERM..ETIMEDOUT) so callers don't pull <errno.h>.
  - Misra/Std/Math.h: F64Abs / F64IsNan / F64IsInf / F64Pow via
    compiler builtins (no <math.h>).
  - Types.h: hex INT*_MIN/MAX, hex UINT*_MAX, MAX_ALIGN, offsetof
    macro -- no <limits.h>/<stdint.h>/<stddef.h>.
  - .gitignore: __pycache__ / *.pyc.

Tests: Linux 104/104, Mac 99/99 (5 Linux-only platform tests skipped
as expected). check_no_libc clean. clang-format applied.

+++ Source/Misra/Std/Allocator/Arena.c
@@ -9,8 +9,6 @@
 #include <Misra/Std/Log.h>
 #include <Misra/Std/Memory.h>
 
-#include <stdint.h>
-
 // Relational invariants for ArenaAllocator:
 //   - head and tail chain endpoints are both NULL or both non-NULL.
 //   - last_ptr / last_size record the most recent allocation for
@@ -111,10 +109,10 @@ void *arena_allocator_allocate(Allocator *self, size bytes, i8 zeroed) {
 
     ArenaChunk *chunk = arena->tail;
     if (chunk) {
-        uintptr_t base_addr    = (uintptr_t)chunk->base;
-        uintptr_t free_addr    = base_addr + chunk->used;
-        uintptr_t aligned_addr = (free_addr + (uintptr_t)(align - 1)) & ~(uintptr_t)(align - 1);
-        size      aligned_used = (size)(aligned_addr - base_addr);
+        u64  base_addr    = (u64)chunk->base;
+        u64  free_addr    = base_addr + chunk->used;
+        u64  aligned_addr = (free_addr + (u64)(align - 1)) & ~(u64)(align - 1);
+        size aligned_used = (size)(aligned_addr - base_addr);
         if (aligned_used + padded <= chunk->capacity) {
             char *result     = chunk->base + aligned_used;
             chunk->used      = aligned_used + padded;
@@ -135,13 +133,13 @@ void *arena_allocator_allocate(Allocator *self, size bytes, i8 zeroed) {
     }
     arena->tail = chunk;
 
-    uintptr_t base_addr    = (uintptr_t)chunk->base;
-    uintptr_t aligned_addr = (base_addr + (uintptr_t)(align - 1)) & ~(uintptr_t)(align - 1);
-    size      aligned_used = (size)(aligned_addr - base_addr);
-    char     *result       = chunk->base + aligned_used;
-    chunk->used            = aligned_used + padded;
-    arena->last_ptr        = result;
-    arena->last_size       = padded;
+    u64   base_addr    = (u64)chunk->base;
+    u64   aligned_addr = (base_addr + (u64)(align - 1)) & ~(u64)(align - 1);
+    size  aligned_used = (size)(aligned_addr - base_addr);
+    char *result       = chunk->base + aligned_used;
+    chunk->used        = aligned_used + padded;
+    arena->last_ptr    = result;
+    arena->last_size   = padded;
     return result;
 }
